extract common utilities to a new header
